Update LatLngUTMConverter.cs
First of all: awesome library. Converts Lat/Lon to UTM without issues but not so well when converting from UTM to Lat/Lon because the latitude will almost always be negative. The issues seems to be here: 
if ("N" == ZoneLetter)
            {
                NorthernHemisphere = 1;
            }
            else
            {
                NorthernHemisphere = 0;
                y -= 10000000.0;
}

So when ZoneLetter is equal to "N", 1 gets assigned to NorthernHemisphere variable but NorthernHemisphere isn't use anywhere. Second problem, when ZoneLetter is NOT equal to "N", Latitude ends up being a negative number which doesn't translate to correct coordinates. I suggest completely removing this part of the code because without it works just fine.